requirements.txt
streamlit
faiss-cpu
sentence-transformers
llama-cpp-python==0.2.20
numpy
pandas
requests
feedparser
newspaper3k
langchain